ring loadables v6.41

+++ ring_loadables/test.sh
@@ -66,7 +66,8 @@ spawn_worker() {
 		echo "$CNT" >> ./count.${1}
                 ((ITER++))
                 mapfile -t -u $fd_read -n $CNT A
-                : "${A[@]}"
+                (( ${#A[@]} == CNT )) || echo "ERROR on iteration $ITER: expected $CNT values, got ${#A[@]} values" >&2
+                : ${A[@]}
             done
             ring_worker dec  # de-register worker
 	    echo "$total" >./total.${1}
@@ -75,6 +76,135 @@ spawn_worker() {
         } {fd_read}<test.dat
     ) &
     P+=($!)
+}
+;;
+
+3)
+spawn_worker() {
+    (
+        {
+            ITER=0
+            ring_worker inc  # Register ourselves as 1 worker
+            IFS=' '
+            while ring_claim OFF CNT $fd_read; do
+                [[ "$CNT" == "0" ]] && break
+                ((total+=CNT))
+                echo "$total" >./total.${1}
+		echo "$CNT" >> ./count.${1}
+                ((ITER++))
+                mapfile -t -u $fd_read -n $CNT A
+                (( ${#A[@]} == CNT )) || echo "ERROR on iteration $ITER: expected $CNT values, got ${#A[@]} values" >&2
+                 : ${A[*]}
+                 echo
+            done
+            ring_worker dec  # de-register worker
+	    echo "$total" >./total.${1}
+            printf 'TOTAL=%s    ITER=%s    AVG=%s\n' "$total" "$ITER" "$((total/ITER))" >./final.${1}
+            
+        } {fd_read}<test.dat
+    ) &
+    P+=($!)
+}
+;;
+
+4)
+spawn_worker() {
+    (
+        {
+            ITER=0
+            ring_worker inc  # Register ourselves as 1 worker
+            while ring_claim OFF CNT $fd_read; do
+                [[ "$CNT" == "0" ]] && break
+                ((total+=CNT))
+                echo "$total" >./total.${1}
+		echo "$CNT" >> ./count.${1}
+                ((ITER++))
+                mapfile -t -u $fd_read -n $CNT A
+                (( ${#A[@]} == CNT )) || echo "ERROR on iteration $ITER: expected $CNT values, got ${#A[@]} values" >&2
+                printf '%s\n' "${A[@]}"
+            done | wc -l >>./count.out.${1}
+            ring_worker dec  # de-register worker
+	    echo "$total" >./total.${1}
+            printf 'TOTAL=%s    ITER=%s    AVG=%s\n' "$total" "$ITER" "$((total/ITER))" >./final.${1}
+            
+        } {fd_read}<test.dat
+    ) &
+    P+=($!)
+}
+;;
+
+5)
+spawn_worker() {
+    (
+        {
+            ITER=0
+            ring_worker inc  # Register ourselves as 1 worker
+            while ring_claim OFF CNT $fd_read; do
+                [[ "$CNT" == "0" ]] && break
+                ((total+=CNT))
+                echo "$total" >./total.${1}
+		echo "$CNT" >> ./count.${1}
+                ((ITER++))
+                mapfile -t -u $fd_read -n $CNT A
+                (( ${#A[@]} == CNT )) || echo "ERROR on iteration $ITER: expected $CNT values, got ${#A[@]} values" >&2
+		IFS=' '
+                printf '%s\n' ${A[*]@Q}
+            done | wc -l >>./count.out.${1}
+            ring_worker dec  # de-register worker
+	    echo "$total" >./total.${1}
+            printf 'TOTAL=%s    ITER=%s    AVG=%s\n' "$total" "$ITER" "$((total/ITER))" >./final.${1}
+            
+        } {fd_read}<test.dat
+    ) &
+    P+=($!)
+}
+;;
+
+6)
+
+ff() {
+sha1sum "${@}"
+sha256sum "${@}"
+sha512sum "${@}"
+sha224sum "${@}"
+sha384sum "${@}"
+md5sum "${@}"
+sum -s "${@}"
+sum -r "${@}"
+cksum "${@}"
+b2sum "${@}"
+cksum -a sm3 "${@}"
+xxhsum "${@}"
+xxhsum -H3 "${@}"
+}
+export -f ff
+
+spawn_worker() {
+  {
+    (
+        {
+            ITER=0
+            ring_worker inc  # Register ourselves as 1 worker
+            exec {fd_count}<><(:)
+            #{ cat <&$fd_count | wc -l > ./count.final.${1} } &
+            while ring_claim OFF CNT $fd_read; do
+                [[ "$CNT" == "0" ]] && break
+                ((total+=CNT))
+                echo "$total" >./total.${1}
+                echo "$CNT" >> ./count.${1}
+                ((ITER++))
+                mapfile -t -u $fd_read -n $CNT A
+                ff "${A[@]}"
+            done
+            ring_worker dec  # de-register worker
+            #exec {fd_count}>&-
+	    echo "$total" >./total.${1}
+            printf 'TOTAL=%s    ITER=%s    AVG=%s\n' "$total" "$ITER" "$((total/ITER))" >./final.${1}
+
+        } {fd_read}</mnt/ramdisk/flist 1>&${fd1} 2>&${fd2}
+    ) &
+    P+=($!)
+  } {fd1}>&1 {fd2}>&2
 }
 ;;
 
@@ -103,30 +233,55 @@ spawn_worker() {
 
 esac
 
+
+
+[[ $(echo ./total.* ./count.* ./final.* ) ]] && \rm ./total.* ./count.* ./final.*
+
+case "$1" in
+6)
+exec {fd_scan}</mnt/ramdisk/flist
+dataN="$(wc -l /mnt/ramdisk/flist)"
+;;
+*)
+
 dataN=1000000000
 
-echo "Generating test data..."
-yes $'\n' | head -n $dataN > test.dat
-
-\rm ./total.* ./count.* ./final.*
-
+echo "Generating test data..." >&2
+case "$1" in
+	4|5)
+		seq $dataN >test.dat
+	;;
+	*)
+		yes $'\n' | head -n $dataN > test.dat
+	;;
+esac
 exec {fd_scan}<test.dat
+;;
+esac
+
+
 exec {fd_spawn}<><(:)
+total=0
+P=()
+
+export nWorkers=1
+export nWorkersMax=$(( 1 * $(nproc) ))
+export nBytesMax=ARG_MAX
+export nLinesMax=1024
+export nBatchMax=1024
+
+start_time=${EPOCHREALTIME//./}
 
 ring_init
 
-echo "Starting scanner..."
+echo "Starting scanner..." >&2
 ( ring_scanner ${fd_scan} ${fd_spawn} ) &
 SCANNER_PID=$!
 
 # 4. Consumer Loop
-echo "Consuming..."
-total=0
-P=()
-export nWorkers=1
-export nWorkersMax=$(( 1 * $(nproc) ))
+echo "Consuming..." >&2
 
-start_time=${EPOCHREALTIME//./}
+
 sleep 0.1s
 for (( nn=0; nn<nWorkers; nn++)); do
     spawn_worker "$nn"
@@ -134,7 +289,8 @@ done
 
 while true; do
     read -r -u $fd_spawn N
-    [[ "$N" == 'x' ]] && { printf '\nSCANNER HIT EOF\nelapsed time = %s us\n' $(( ${EPOCHREALTIME//./} - start_time )); break; }
+    [[ "$N" == 'x' ]] && { printf '\nSCANNER HIT EOF\nelapsed time = %s us\n' $(( ${EPOCHREALTIME//./} - start_time )) >&2; break; }
+    echo "spawning $N workers" >&2
     nWorkers0="$nWorkers"
     (( ( nWorkers0 + N ) > nWorkersMax )) && (( N = nWorkersMax - nWorkers0 ))
     (( N > 0 )) && for ((nn=nWorkers0; nn<nWorkers0+N; nn++)); do
@@ -153,9 +309,9 @@ for nn in ./total.*; do
 (( total = total + total0 ))
 done
 
-echo "Done."
-echo "Total Lines Consumed: $total (Expected: $dataN)"
-echo "Time: ${elapsed}us"
+echo "Done." >&2
+echo "Total Lines Consumed: $total (Expected: $dataN)" >&2
+echo "Time: ${elapsed} us" >&2
 
 # Cleanup
 ring_destroy
